+++ src/domain/material-movimentation/application/use-cases/movimentation/transfer-movimentation-between-projects.ts
@@ -281,7 +281,10 @@ export class TransferMovimentationBetweenProjectsUseCase {
         new Date(
           Math.max(
             ...movimentations
-              .filter((movimentation) => (movimentation.projectId = project.id))
+              .filter(
+                (movimentation) =>
+                  movimentation.projectId.toString() === project.id.toString()
+              )
               .map((movimentation) => movimentation.createdAt.getTime())
           )
         ).setMilliseconds(999)
@@ -291,7 +294,10 @@ export class TransferMovimentationBetweenProjectsUseCase {
         new Date(
           Math.min(
             ...movimentations
-              .filter((movimentation) => (movimentation.projectId = project.id))
+              .filter(
+                (movimentation) =>
+                  movimentation.projectId.toString() === project.id.toString()
+              )
               .map((movimentation) => movimentation.createdAt.getTime())
           )
         ).setMilliseconds(0)